Run reframe tests after building a uenv (#14)
This enables the test stage through a firecrest-submitted job (exactly like the build stage).

This allocates a new job on a compute node to submit the reframe tests. This is overkill and could be changed to use a baremetal (login node) runner in the future, or to use a single-gpu allocation when node sharing is allowed on all clusters. Until then, this is likely sufficient and is the same setup that is used in cscs-reframe-tests pipelines.